+++ index.js
@@ -3,10 +3,8 @@ var bodyParser = require('body-parser');
 var express = require('express');
 var gitHubWebHook = require('express-github-webhook');
 var nconf = require('nconf');
-var Promise = require('bluebird');
 
-var GitHubServer = require('./lib/GitHubServer');
-var getGoogleGroupLinks = require('./lib/getGoogleGroupLinks');
+var commentOnClosedIssue = require('./lib/commentOnClosedIssue');
 
 var app = express();
 module.exports = app;
@@ -20,7 +18,6 @@ nconf.defaults({
     port: 5000,
     secret: '', // Repository secret to verify incoming WebHook requests from GitHub
     gitHubToken: '', // Token used to verify outgoing requests to GitHub repository
-    repository: '', // Repository to scan for outdated pull requests and bump them
     listenPath: '/' // Path on which to listen for incoming requests
 });
 
@@ -32,35 +29,13 @@ var webHookHandler = gitHubWebHook({
 app.use(bodyParser.json());
 app.use(webHookHandler);
 
-var gitHubServer = new GitHubServer('cesium-concierge', nconf.get('gitHubToken'));
-
-/** Get comments -> regex search -> post comment
- *
- * @param {String} commentsUrl URL to set/get comments on issue (https://developer.github.com/v3/activity/events/types/#issuesevent)
- * @return {Promise<http.IncomingMessage>} Response
- */
-function commentOnClosedIssue(commentsUrl) {
-    var comments = [];
-    var linkMatches = [];
-
-    return gitHubServer.get(commentsUrl)
-    .then(function(commentsJsonResponse) {
-        comments = GitHubServer.getCommentsFromResponse(commentsJsonResponse);
-        linkMatches = getGoogleGroupLinks(comments);
-        if (linkMatches === []) {
-            return Promise.reject('No google group links found in comments!');
-        }
-        console.log('Found these links in the comments: ', linkMatches);
-        return gitHubServer.postComment(commentsUrl,
-            'Please make sure to update ' + linkMatches + ' on this closed issue.\n\n__I am a bot BEEEP BOOOP__');
-    });
-}
-
 webHookHandler.on('issues', function(repo, jsonResponse) { // eslint-disable-line no-unused-vars
-    var commentsUrl = GitHubServer.issue.getCommentsUrl(jsonResponse);
     switch (jsonResponse.action) {
         case 'closed':
-            commentOnClosedIssue(commentsUrl).then(function(status) {
+            commentOnClosedIssue(jsonResponse, {
+                'User-Agent': 'cesium-concierge',
+                Authorization: 'token ' + nconf.get('gitHubToken')
+            }).then(function(status) {
                 console.log('GitHub API returned with:', status);
             }).catch(function(e) {
                 console.log('commentOnClosedIssue got an error:', e);
